Fix graphs
