+++ OV_DB/Models/Graphs/Dataset.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 namespace OV_DB.Models.Graphs
@@ -9,12 +10,15 @@ namespace OV_DB.Models.Graphs
     {
         public string Label { get; set; }
         public List<Point> Data { get; set; }
-        public bool SteppedLine { get; set; } = true;
-        public double CategoryPercentage = 0.7d;
+        public bool Stepped { get; set; } = true;
+        public double CategoryPercentage = 1.0d;
         public double BarPercentage { get; set; } = 1.0d;
         public string BarThickness = "flex";
-        public string Stack { get; set; }
+        public bool Stack { get; set; }
         public string BorderColor { get; set; }
-        public string PointBackgroundColor { get; set; }
+        public bool Fill { get; set; }
+        public string BackgroundColor { get; set; }
+        [JsonPropertyName("yAxisID")]
+        public string YAxisID = "y";
     }
 }
