+++ OV_DB/OVDBFrontend/angular.json
@@ -8,7 +8,7 @@
       "schematics": {
         "@schematics/angular:component": {
           "style": "scss",
-          "standalone": false
+          "standalone": true
         }
       },
       "root": "",
