+++ OV_DB/OVDBFrontend/package.json
@@ -31,10 +31,12 @@
     "@ngx-translate/core": "^16.0.0",
     "@ngx-translate/http-loader": "^16.0.0",
     "chart.js": "^4.4.0",
+    "chartjs-adapter-luxon": "^1.3.1",
     "chartjs-plugin-zoom": "^2.0.1",
     "file-saver": "^2.0.5",
     "leaflet": "^1.9.4",
     "leaflet.markercluster": "^1.5.3",
+    "luxon": "^3.6.1",
     "moment": "^2.29.4",
     "ng2-charts": "^6.0.0",
     "rxjs": "^6.6.7",
