+++ OV_DB/OVDBFrontend/src/app/stats/stats.component.html
@@ -1,60 +1,13 @@
-<div class="container">
-  <mat-card appearance="outlined">
-    <mat-card-title>{{'SELECTION'|translate}}</mat-card-title>
-    <mat-form-field>
-      <mat-label>{{'CHOOSEMAP'|translate}}</mat-label>
-      <mat-select (selectionChange)="changeMap($event.value)">
-        @for (map of maps; track map) {
-          <mat-option [value]="map.mapGuid">
-            {{map.name}}
-          </mat-option>
-        }
-      </mat-select>
-    </mat-form-field>
-    <mat-form-field>
-      <mat-label>{{'CHOOSEYEAR'|translate}}</mat-label>
-      <mat-select (selectionChange)="getData($event.value)" [(ngModel)]="selectedYear">
-        @for (year of years; track year) {
-          <mat-option [value]="year">
-            {{year}}
-          </mat-option>
-        }
-        <mat-option [value]="null">
-          {{'ALL'|translate}}
-        </mat-option>
-      </mat-select>
-    </mat-form-field>
-  </mat-card>
-
-  <div class="row">
-    @if (!!tableData) {
-      <table>
-        <tr>
-          <th>{{'NAME'|translate}}</th>
-          <th>{{'DISTANCE'|translate}}</th>
-        </tr>
-        @for (row of tableData; track row) {
-          <tr>
-            <td>{{name(row)}}</td>
-            <td>{{row.distance}} km</td>
-          </tr>
-        }
-        <button mat-raised-button (click)="download()">{{'TRIPREPORT.DOWNLOAD'|translate}}</button>
-        <button mat-raised-button (click)="export()">{{'TRIPREPORT.EXPORT'|translate}}</button>
-      </table>
+<mat-tab-group>
+  <mat-tab [label]="'HOME.STATS'|translate">
+    <app-time-stats />
+  </mat-tab>
+  <mat-tab [label]="'STATS.OPERATORS'|translate">
+    @defer (on viewport) {
+    <app-used-operators />
     }
-    <!-- <div *ngIf="!!data" class="container graph">
-    <canvas baseChart height="300" [datasets]="data.datasets" [options]="lineChartOptions" chartType="line"></canvas>
-  </div> -->
-</div>
-<div class="map-container" #mapContainer>
-  <div class="map" leaflet [leafletOptions]="options" [leafletLayers]="layers"
-    [leafletLayersControl]="leafletLayersControl" [leafletFitBounds]="bounds">
-  </div>
-  <div [ngClass]="{'loading':loadingMap,'loading-indicator':true}">
-    <mat-spinner></mat-spinner>
-  </div>
-</div>
-
-
-</div>
+    @placeholder {
+    <p>loading</p>
+    }
+  </mat-tab>
+</mat-tab-group>
