+++ OV_DB/OVDBFrontend/src/app/stats/time-stats/time-stats.component.html
@@ -1,0 +1,77 @@
+<div class="container">
+    <mat-card appearance="outlined">
+        <mat-card-title>{{'SELECTION'|translate}}</mat-card-title>
+        <mat-form-field>
+            <mat-label>{{'CHOOSEMAP'|translate}}</mat-label>
+            <mat-select (selectionChange)="changeMap($event.value)">
+                @for (map of maps; track map) {
+                <mat-option [value]="map.mapGuid">
+                    {{map.name}}
+                </mat-option>
+                }
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>{{'CHOOSEYEAR'|translate}}</mat-label>
+            <mat-select (selectionChange)="getData($event.value)" [(ngModel)]="selectedYear">
+                @for (year of years; track year) {
+                <mat-option [value]="year">
+                    {{year}}
+                </mat-option>
+                }
+                <mat-option [value]="0">
+                    {{'ALL'|translate}}
+                </mat-option>
+            </mat-select>
+        </mat-form-field>
+    </mat-card>
+
+    <div class="row">
+        @if (!!tableData) {
+        <table>
+            <tr>
+                <th>{{'NAME'|translate}}</th>
+                <th>{{'DISTANCE'|translate}}</th>
+            </tr>
+            @for (row of tableData; track row) {
+            <tr>
+                <td>{{name(row)}}</td>
+                <td>{{row.distance}} km</td>
+            </tr>
+            }
+            <button mat-raised-button (click)="download()">{{'TRIPREPORT.DOWNLOAD'|translate}}</button>
+            <button mat-raised-button (click)="export()">{{'TRIPREPORT.EXPORT'|translate}}</button>
+        </table>
+        }
+        @if(!!data || !!singleData){
+        <mat-tab-group class="graph-tab">
+            <mat-tab style="max-height: 500px;" [label]="'STATS.CUMULATIVE'|translate">
+                @if(!!data){
+                <div class="container graph">
+                    <canvas #cumulativeChart baseChart [data]="data" [options]="lineChartOptions"
+                        [type]="'line'"></canvas>
+                </div>
+                }
+            </mat-tab>
+            <mat-tab style="max-height: 500px;" [label]="'STATS.PER_DAY'|translate">
+                @if(!!singleData){
+                <div class="container graph">
+                    <canvas #singleChart baseChart [data]="singleData" [options]="barChartOptions"
+                        [type]="'bar'"></canvas>
+                </div>
+                }
+            </mat-tab>
+        </mat-tab-group>
+        }
+    </div>
+    <div class="map-container" #mapContainer>
+        <div class="map" leaflet [leafletOptions]="options" [leafletLayers]="layers"
+            [leafletLayersControl]="leafletLayersControl" [leafletFitBounds]="bounds">
+        </div>
+        <div [ngClass]="{'loading':loadingMap,'loading-indicator':true}">
+            <mat-spinner></mat-spinner>
+        </div>
+    </div>
+
+
+</div>
