+++ OV_DB/OVDBFrontend/src/assets/i18n/en.json
@@ -218,5 +218,10 @@
     "REQUESTS": "Requests",
     "MAPS": "Maps",
     "DATA": "Data"
+  },
+  "STATS": {
+    "OPERATORS": "Operators",
+    "CUMULATIVE": "Cumulative",
+    "PER_DAY": "Per day"
   }
 }
